// leave-service/src/models/constants.ts
export enum LeaveType {
  PRIVILEGE = "PRIVILEGE",
  SICK = "SICK",
  CASUAL = "CASUAL",
}

export enum DefaultLeaveAllocation {
  CASUAL = 12,
  SICK = 10,
  PRIVILEGE = 15,
}

export enum LeaveStatus {
  PENDING = "PENDING",
  APPROVED = "APPROVED",
  REJECTED = "REJECTED",
  CANCELLED = "CANCELLED",
}